+++ src/EntityGraphQL/Schema/FieldExtensions/Filter/FilterExpressionExtension.cs
@@ -37,18 +37,18 @@ public class FilterExpressionExtension : BaseFieldExtension
 
     public override (Expression? expression, ParameterExpression? originalArgParam, ParameterExpression? newArgParam, object? argumentValue) GetExpressionAndArguments(
         IField field,
-        BaseGraphQLField fieldNode,
-        Expression expression,
-        ParameterExpression? argumentParam,
-        dynamic? arguments,
-        Expression context,
-        bool servicesPass,
-        bool withoutServiceFields,
-        ParameterReplacer parameterReplacer,
-        ParameterExpression? originalArgParam,
-        CompileContext compileContext
+        FieldExtensionExpressionContext context
     )
     {
+        var fieldNode = context.FieldNode;
+        var expression = context.Expression;
+        var argumentParam = context.ArgumentParameter;
+        var arguments = context.Arguments;
+        var servicesPass = context.ServicesPass;
+        var parameterReplacer = context.ParameterReplacer;
+        var originalArgParam = context.OriginalArgumentParameter;
+        var compileContext = context.CompileContext;
+
         var filter = arguments?.Filter as EntityQueryType;
         if (arguments != null && filter != null && filter?.HasValue)
         {